// frontend/src/types/setting.ts
export interface Setting {

    id: number;

    /*
    |--------------------------------------------------------------------------
    | Company
    |--------------------------------------------------------------------------
    */

    company_name: string;

    company_email: string;

    company_phone: string | null;

    company_address: string | null;

    about_company: string | null;

    /*
    |--------------------------------------------------------------------------
    | Branding
    |--------------------------------------------------------------------------
    */

    logo: string | null;

    favicon: string | null;

    /*
    |--------------------------------------------------------------------------
    | Social Media
    |--------------------------------------------------------------------------
    */

    facebook: string | null;

    instagram: string | null;

    linkedin: string | null;

    twitter: string | null;

    whatsapp: string | null;

    /*
    |--------------------------------------------------------------------------
    | Website
    |--------------------------------------------------------------------------
    */

    google_map: string | null;

    working_hours: string | null;

    /*
    |--------------------------------------------------------------------------
    | SEO
    |--------------------------------------------------------------------------
    */

    meta_title: string | null;

    meta_description: string | null;

    meta_keywords: string | null;

    /*
    |--------------------------------------------------------------------------
    | System
    |--------------------------------------------------------------------------
    */

    default_language: "fa" | "en";

    copyright: string | null;

    created_at: string;

    updated_at: string;

}